requirements.txt
alembic==1.6.0
psycopg2==2.9.1
Flask==2.0.2
PyJWT==2.3.0
flask_cors==3.0.10
